Fix 1

+++ zkozz/src/main/java/com/leo/zkozz/model/repository/FilialRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FilialRepository extends JpaRepository<Filial, Integer> {
+public interface FilialRepository extends JpaRepository<Filial, Long> {
 
 }
